backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
# Pin httpx to a version compatible with openai 1.3.x to avoid 'proxies' kw error
httpx==0.27.2
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
cors==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
